+++ AutoBar/AutoBarBar/Models/Product.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public double UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
         public string ImageLink { get; set; }
     }
 }
